refactor(sandbox): own the isolated-vm hook executor
The Worker classifies inbound channel deliveries in-process, and it may not
import the API, so the isolate could not stay in apps/api/src/hooks. It moves to
@tulipfarm/sandbox whole — analyzer, executor, worker host, and protocol — and
the API's hook routes now call the package. One sandbox implementation, one
place where untrusted code is bounded.
