+++ packages/sandbox/package.json
@@ -8,8 +8,14 @@
     "test": "vitest run --passWithNoTests",
     "typecheck": "tsc --noEmit"
   },
+  "dependencies": {
+    "isolated-vm": "^7.0.0"
+  },
   "devDependencies": {
     "@tulipfarm/tsconfig": "workspace:*",
-    "typescript": "^6.0.3"
+    "@types/node": "^26.1.1",
+    "tsx": "^4.22.4",
+    "typescript": "^6.0.3",
+    "vitest": "^4.1.9"
   }
 }
