+++ packages/sandbox/src/hooks/analyzer.ts
@@ -2,6 +2,13 @@ export class HookAnalysisError extends Error {
   name = "HookAnalysisError";
 }
 
+/**
+ * Source patterns refused before a hook is ever compiled.
+ *
+ * This is a cheap first filter, not the isolation boundary — the isolate is. It exists so an
+ * obvious escape attempt is rejected with a readable reason instead of failing opaquely inside a
+ * sandbox that was never going to grant it anyway.
+ */
 const BANNED: RegExp[] = [
   /\brequire\s*\(/,
   /\bimport\s*\(/,
